userAuthmysql solution

+++ userAuth/php/userauth.php
@@ -1,14 +1,26 @@
 <?php
 
-require "../config.php";
+require_once "../config.php";
 
 //register users
-function registerUser($username, $email, $password){
+function registerUser($fullnames, $email, $password, $gender, $country){
     //create a connection variable using the db function in config.php
     $conn = db();
 
-    echo "SAVE MY INFO TO THE DATABASE :) (IMPLEMENT ME)";
-   //check if user with this email already exist in the database
+    if(mysqli_num_rows(mysqli_query($conn, "SELECT * FROM Students WHERE email='$email'"))>=1){
+        echo "<script> alert('User Email Already Taken')</script>";
+        header("refresh: 0.5; url=../forms/register.html");
+    }
+    else{
+        $sql = "INSERT INTO `Students` (`full_names`, `country`, `email`, `gender`, `password`) VALUES ('$fullnames', '$country', '$email', '$gender', '$password')";
+        if(mysqli_query($conn, $sql)){
+            echo "<script> alert('User Successfully created!!')</script>";
+            header("refresh: 2; url=../forms/login.html");
+        }
+        else{
+            echo "<script'> alert('An Error Occured please try again')</script>";
+        }
+    }
 }
 
 
@@ -16,33 +28,51 @@ function registerUser($username, $email, $password){
 function loginUser($email, $password){
     //create a connection variable using the db function in config.php
     $conn = db();
-
-    echo "LOG ME IN (IMPLEMENT ME)";
-    //open connection to the database and check if username exist in the database
-    //if it does, check if the password is the same with what is given
-    //if true then set user session for the user and redirect to the dasbboard
+    $query = "SELECT * FROM Students WHERE email='$email' AND password='$password'";
+    $result = mysqli_query($conn, $query);
+    if(mysqli_num_rows($result) >= 1){
+        session_start();
+        $_SESSION['username'] = $email;
+        header("Location: ../dashboard.php");
+        }
+    else{
+        echo "<script> alert('Incorrect Password') </script>";
+    }
 }
 
 
 function resetPassword($email, $password){
     //create a connection variable using the db function in config.php
     $conn = db();
-    echo "RESET YOUR PASSWORD (IMPLEMENT ME)";
-    //open connection to the database and check if username exist in the database
-    //if it does, replace the password with $password given
+    if(mysqli_num_rows(mysqli_query($conn, "SELECT * FROM Students WHERE email='$email'"))>=1){
+        $sql = "UPDATE table Student set password='$password' Where email='$email'";
+        if(mysqli_query($conn, $sql)){
+            echo "<script> alert('Password Succesfully updated!!')</script>";
+        }
+        else{
+            echo "<script> alert('An Error Occured please try again')</script>";
+        }
+    }
+    else{
+        echo "an error occured";
+    }
 }
 
 function getusers(){
     $conn = db();
     $sql = "SELECT * FROM Students";
     $result = mysqli_query($conn, $sql);
+    echo"<html>
+    <head></head>
+    <body>
+    <center><h1><u> ZURI PHP STUDENTS </u> </h1> 
+    <table border='1' style='width: 700px; background-color: magenta; border-style: none'; >
+    <tr style='height: 40px'><th>ID</th><th>Full Names</th> <th>Email</th> <th>Gender</th> <th>Country</th> <th>Action</th></tr>";
     if(mysqli_num_rows($result) > 0){
         while($data = mysqli_fetch_assoc($result)){
             //show data
-            echo "<html>
-            <head></head>
-            <body>
-                <table border='1' style='width: 700px;'><tr><td style='width: 50px'>" . $data['id'] . "</td>
+            echo "<tr style='height: 30px'>".
+                "<td style='width: 50px; background: blue'>" . $data['id'] . "</td>
                 <td style='width: 150px'>" . $data['full_names'] .
                 "</td> <td style='width: 150px'>" . $data['email'] .
                 "</td> <td style='width: 150px'>" . $data['gender'] . 
@@ -52,24 +82,35 @@ function getusers(){
                 <input type='hidden' name='id'" .
                  "value=" . $data['id'] . ">".
                 "<td style='width: 150px'> <button type='submit', name='delete'> DELETE </button>".
-                "</tr></table>
-                </table>
-            </body>
-            </html>";
+                "</tr>";
         }
+        echo "</table></table></center></body></html>";
     }
     //return users from the database
     //loop through the users and display them on a table
 }
 
- function deleteaccount($conn, $id){
+function logout(){
+    if($_SESSION['username'])
+    {
+        session_unset();
+        session_destroy();
+        header("Location: ../forms/login.html");
+    }
+    else{
+        header("Location: ../forms/login.html");
+    }
+    
+}
+
+ function deleteaccount($id){
      $conn = db();
     //reference how i used the connection here, to help you with the others
-     $sql = "DELETE FROM Students WHERE id=$id";
-     if(mysqli_query($conn, $sql)){
-         echo "<h1 style='color: green'>DELETED</h1>";
+    if(mysqli_num_rows(mysqli_query($conn, "SELECT * FROM Students WHERE id=$id"))){
+        $sql = "DELETE FROM Students WHERE id=$id";
+        if(mysqli_query($conn, $sql)){
+            echo "<script> alert('DELETED')</script>";
+            header("refresh:0.5; url=action.php?all=");
      }
-     //echo "DELETE YOUR ACCOUNT (IMPLEMENT ME)";
-
-     //delete user with the given id from the database
+    }
  }
